requirements.txt
python-fasthtml
google-genai
Pillow
PyMuPDF
python-multipart
